requirements.txt
asgiref
autopep8
Django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
numpy
opencv-contrib-python
opencv-python
Pillow
crispy_forms